modifié :         copycd.sh

+++ scripts/copycd.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 cdname=$1
+clean=$2
 
 cddev=$(mount |grep -i $cdname |cut -d " " -f 1)
 
@@ -20,4 +21,8 @@ echo "./cdrdao read-cd --datafile $cdname.bin --driver generic-mmc:0x00020000 --
 ./cdrdao read-cd --datafile $cdname.bin --driver generic-mmc:0x00020000 --device $cddev $cdname.toc
 
 ./toc2cue $cdname.toc $cdname.cue
+
+if [ $clean == 1 ]; then
+echo "clean est égal à $clean"
 rm -f -r $cdname.toc versions.txt toc2cue cdrdao
+fi
